Unified the cursor animation code
